+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for NeoFS Node
 ### Added
 - Indexes inspection command to neofs-lens (#2882)
 - Add objects sanity checker to neofs-lens (#2506)
+- Support for 0.20.0+ neofs-contract archive format (#2872)
 
 ### Fixed
 - Control service's Drop call does not clean metabase (#2822)